Adding all the outstanding quality control issues for the purposes of refactoring
