+++ src/acceptanceTest/java/uk/gov/hmcts/futurehearings/hmi/acceptance/common/security/OAuthTokenGenerator.java
@@ -12,25 +12,25 @@ import org.springframework.http.HttpStatus;
 public class OAuthTokenGenerator {
 
     public static final String generateOAuthToken(final String token_apiURL,
-                                                final String token_apiTenantId,
-                                                final String grantType,
-                                                final String clientID,
-                                                final String clientSecret,
-                                                final String scope,
-                                                final HttpStatus httpStatus) throws Exception {
+                                                  final String token_apiTenantId,
+                                                  final String grantType,
+                                                  final String clientID,
+                                                  final String clientSecret,
+                                                  final String scope,
+                                                  final HttpStatus httpStatus) throws Exception {
 
         String full_token_apiURL = String.format(token_apiURL, token_apiTenantId);
         final String bodyForToken = String.format("grant_type=%s&client_id=%s&client_secret=%s&scope=%s",
                 grantType, clientID, clientSecret, scope);
 
         Response response = callTokenGeneratorEndpoint(bodyForToken, httpStatus, full_token_apiURL);
-        assertEquals(httpStatus.value(),response.getStatusCode());
+        assertEquals(httpStatus.value(), response.getStatusCode());
         return response.jsonPath().getString("access_token");
     }
 
     public static final Response callTokenGeneratorEndpoint(final String bodyForToken,
-                                                         final HttpStatus badRequest,
-                                                         final String full_token_apiURL) {
+                                                            final HttpStatus badRequest,
+                                                            final String full_token_apiURL) {
 
         log.debug("The value of the Body : " + bodyForToken);
         log.debug("The value of the Target URL : " + full_token_apiURL);
