+++ src/smokeTest/java/uk/gov/hmcts/futurehearings/hmi/smoke/HmiApiSmokeTest.java
@@ -24,6 +24,7 @@ import org.springframework.test.context.ActiveProfiles;
 @SpringBootTest(classes = {Application.class})
 @ActiveProfiles("smoke")
 @DisplayName("Smoke Test for the Root HMI Context")
+@SuppressWarnings("java:S2187")
 class HmiApiSmokeTest extends SmokeTest {
 
     @Value("${hmiApiRootContext}")
