+++ src/smokeTest/java/uk/gov/hmcts/futurehearings/hmi/smoke/notifications/NotificationsApiSmokeTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.futurehearings.hmi.smoke.notification;
+package uk.gov.hmcts.futurehearings.hmi.smoke.notifications;
 
 import static io.restassured.RestAssured.given;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -6,7 +6,6 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import uk.gov.hmcts.futurehearings.hmi.Application;
 import uk.gov.hmcts.futurehearings.hmi.smoke.common.test.SmokeTest;
 
-import io.restassured.response.Response;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.Setter;
@@ -14,10 +13,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.HttpStatus;
 import org.springframework.test.context.ActiveProfiles;
 
 @Setter(AccessLevel.PUBLIC)
@@ -28,7 +25,7 @@ import org.springframework.test.context.ActiveProfiles;
 @DisplayName("Smoke Test for the HMI Notification Context")
 @Disabled("For as the endpoint for Notification API is not available")
 @SuppressWarnings("java:S2187")
-class NotificationApiSmokeTest extends SmokeTest {
+class NotificationsApiSmokeTest extends SmokeTest {
 
     @Value("${notificationApiRootContext}")
     private String notificationApiRootContext;
