+++ src/smokeTest/java/uk/gov/hmcts/futurehearings/hmi/smoke/sessions/SessionsApiSmokeTest.java
@@ -23,6 +23,7 @@ import org.springframework.test.context.ActiveProfiles;
 @SpringBootTest(classes = {Application.class})
 @ActiveProfiles("smoke")
 @DisplayName("Smoke Test for the HMI Sessions Context")
+@SuppressWarnings("java:S2187")
  class SessionsApiSmokeTest extends SmokeTest {
 
     @Value("${sessionsApiRootContext}")
