+++ src/smokeTest/java/uk/gov/hmcts/futurehearings/hmi/smoke/users/UsersApiSmokeTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.futurehearings.hmi.smoke.updateuser;
+package uk.gov.hmcts.futurehearings.hmi.smoke.users;
 
 import static io.restassured.RestAssured.given;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -6,15 +6,12 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import uk.gov.hmcts.futurehearings.hmi.Application;
 import uk.gov.hmcts.futurehearings.hmi.smoke.common.test.SmokeTest;
 
-import io.restassured.response.Response;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.HttpStatus;
 import org.springframework.test.context.ActiveProfiles;
 
 @Slf4j
@@ -22,7 +19,8 @@ import org.springframework.test.context.ActiveProfiles;
 @ActiveProfiles("smoke")
 @DisplayName("Smoke Test for the HMI Users Context")
 @Disabled("For as the endpoint for User API is not available")
-class UserApiSmokeTest extends SmokeTest {
+@SuppressWarnings("java:S2187")
+class UsersApiSmokeTest extends SmokeTest {
 
     @Value("${userApiRootContext}")
     public String userApiRootContext;
